+++ src/consts.ts
@@ -1,7 +1,9 @@
-export const cWowerlayContainer = 'wowerlay-container';
+export const cWowerlayBackground = 'wowerlay-background';
 export const cWowerlay = 'wowerlay';
 export const cWowerlayAnimEnter = 'wowerlay-anim-enter';
 export const cWowerlayAnimLeave = 'wowerlay-anim-leave';
 export const sWowerlayX = '--wowerlay-x';
 export const sWowerlayY = '--wowerlay-y';
 export const sWowerlayZ = '--wowerley-z';
+
+export const scrollbarGap = 15;