+++ src/lib.tsx
@@ -1,7 +1,4 @@
 import './styles/Wowerlay.scss';
-import './styles/WowerlayContainer.scss';
 
 export * from './components/Wowerlay';
-export * from './components/WowerlayContainer';
-
 export { createWowerlay } from './plugin/index';